Refactor evetlistLisener into database
